java italian iter_2

+++ src/main/java/org/636766fe1a6d9265ec017833/636766fe1a6d9265ec017833.java
@@ -11,7 +11,7 @@ public class FileUtils {
      */
     public static void forceDeleteOnExit(File file) throws IOException {
         if (file == null) {
-            throw new NullPointerException("Il file non può essere null");
+            throw new NullPointerException("File must not be null");
         }
 
         if (file.isDirectory()) {
@@ -26,26 +26,23 @@ public class FileUtils {
             return;
         }
 
+        if (!directory.isDirectory()) {
+            throw new IOException("Not a directory: " + directory);
+        }
+
         File[] files = directory.listFiles();
-        if (files != null) {
-            for (File file : files) {
-                if (file.isDirectory()) {
-                    deleteDirectoryOnExit(file);
-                } else {
-                    file.deleteOnExit();
-                }
+        if (files == null) {
+            throw new IOException("Failed to list contents of " + directory);
+        }
+
+        for (File file : files) {
+            if (file.isDirectory()) {
+                deleteDirectoryOnExit(file);
+            } else {
+                file.deleteOnExit();
             }
         }
 
         directory.deleteOnExit();
-    }
-
-    public static void main(String[] args) {
-        try {
-            File file = new File("path/to/your/file_or_directory");
-            forceDeleteOnExit(file);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
 }
